requirements.txt
flask>=3.0
psycopg2-binary>=2.9
werkzeug>=3.0
waitress>=3.0
python-dotenv>=1.0
openpyxl>=3.1
pytest>=8.0
pytest-cov>=5.0
ruff>=0.9
